# src/docs/swagger.yaml
definitions:
  dto.ReqCreateUpdateEntry:
    properties:
      description:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      tag_list:
        items:
          type: integer
        type: array
      time_end:
        type: string
      time_start:
        type: string
    required:
    - time_end
    - time_start
    type: object
...
  dto.RespEntry:
    properties:
      description:
        type: string
      duration:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      tag_list:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      time_end:
        type: string
      time_start:
        type: string
      user_id:
        type: integer
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
  models.Tag:
    properties:
      about:
        type: string
      color:
        type: string
      id:
        type: integer
      name:
        type: string
      userID:
        type: integer
    type: object
...
  pkg.Response:
    properties:
      body: {}
    type: object
info:
  contact: {}
paths:
...
  /entry/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an entry
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: can't find entry with such id
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Delete an entry
      tags:
      - entry
    get:
      consumes:
      - application/json
      description: Get entry by id
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get entry
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespEntry'
              type: object
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Show a post
      tags:
      - entry
  /entry/create:
    post:
      consumes:
      - application/json
      description: Create entry
      parameters:
      - description: entry info
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateEntry'
      produces:
      - application/json
      responses:
        "200":
          description: success update entry
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespEntry'
              type: object
        "400":
          description: bad req
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Create entry
      tags:
      - entry
  /entry/edit:
    post:
      consumes:
      - application/json
      description: Update an entry
      parameters:
      - description: entry info
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateEntry'
      produces:
      - application/json
      responses:
        "200":
          description: success update entry
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespEntry'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Update an entry
      tags:
      - entry
...
  /me/entries:
    get:
      description: Get user entries or get user entries for a day
      parameters:
      - description: day for events
        in: query
        name: day
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success get entries
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/dto.RespEntry'
                  type: array
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get user entries
      tags:
      - entry
...
swagger: "2.0"